+++ backend/app/Core/Backup/Config/services.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 use PaginiumCMS\Core\Backup\Services\BackupManager;
 use PaginiumCMS\Core\Backup\Contracts\BackupInterface;
+use PaginiumCMS\Core\Backup\Services\BackupScheduler;
+use PaginiumCMS\Core\Backup\Commands\RunBackupScheduleCommand;
 use PaginiumCMS\Core\FlatFile\Contracts\FileReaderInterface;
 use PaginiumCMS\Core\FlatFile\Contracts\FileWriterInterface;
 
@@ -11,7 +13,6 @@ use function DI\create;
 use function DI\get;
 
 return [
-    // Mapovanie rozhrania na implementáciu
     BackupInterface::class => create(BackupManager::class)
     ->constructor(
         get(FileReaderInterface::class),
@@ -19,4 +20,8 @@ return [
                   __DIR__ . '/../../../storage/backups',
                   __DIR__ . '/../../../storage/app/content'
     ),
+    BackupScheduler::class => create(BackupScheduler::class)
+        ->constructor(get(BackupInterface::class)),
+    RunBackupScheduleCommand::class => create(RunBackupScheduleCommand::class)
+        ->constructor(get(BackupScheduler::class)),
 ];